requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
boto3
sqlmodel
alembic
jinja2==3.1.2
pydantic_settings
aiomysql
bcrypt
greenlet
email-validator
PyJWT
